+++ example/DataChannel/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,16 +5,16 @@
 import FlutterMacOS
 import Foundation
 
+import ant_media_flutter
 import flutter_webrtc
 import path_provider_foundation
 import path_provider_macos
 import shared_preferences_foundation
-import shared_preferences_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  AntMediaFlutter.register(with: registry.registrar(forPlugin: "AntMediaFlutter"))
   FlutterWebRTCPlugin.register(with: registry.registrar(forPlugin: "FlutterWebRTCPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
-  SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
 }